requirements/base.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
billiard==3.3.0.23
celery==3.1.25
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
decorator==4.0.11
Django==1.11.2
django-celery==3.2.1
django-macros-url==0.3.1
django-rest-auth==0.9.1
django-rest-swagger==2.1.2
django-widget-tweaks==1.4.1
djangorestframework==3.6.3
idna==2.5
ipython==5.5.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.10.2
Jinja2==2.9.6
kombu==3.0.37
MarkupSafe==1.0
oauthlib==2.0.2
openapi-codec==1.3.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.3
ptyprocess==0.5.1
Pygments==2.2.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
simplejson==3.11.1
simplegeneric==0.8.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.21.1
traitlets==4.3.2
wcwidth==0.1.7
wheel==0.24.0